pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dimer"
version = "0.1.0"
description = "Universal Data Source Connector Module with Fallback Strategies"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DiMer Team"},
]
keywords = ["data", "connector", "database", "integration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pyarrow>=12.0.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python[pandas,secure-local-storage]>=3.0.0",
    "snowpark-python>=1.0.0",
]
postgresql = [
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
]
mysql = [
    "PyMySQL>=1.0.0",
    "mysql-connector-python>=8.0.0",
]
bigquery = [
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-bigquery-storage>=2.0.0",
    "db-dtypes>=1.0.0",
]
databricks = [
    "databricks-sql-connector>=2.0.0",
    "databricks-connect>=13.0.0",
]
all = [
    "dimer[snowflake,postgresql,mysql,bigquery,databricks]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dimer-diff = "dimer.cli:main"

[project.urls]
Homepage = "https://github.com/dimer/dimer"
Documentation = "https://dimer.readthedocs.io"
Repository = "https://github.com/dimer/dimer"
"Bug Tracker" = "https://github.com/dimer/dimer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dimer*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=dimer --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["dimer"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]